Update icons and shortcuts.
Use QIcon::fromTheme() instead of hard-coding QRC URLs for icons. This
allows us to use native icons on X11 platforms (e.g. Linux), and
therefore improving the native integration of Watermark on all
platforms. For Windows and macOS, the icons8 style was switched to
Office Small, almost because of personal preferences, but also because
Office theme is more "professional" than the previously "color" theme.