Merge AboutDialog and LicenseDialog.
Most of the software include the license information into the about
dialog, so why not we? Moreover, since previous commits, an "About Qt"
dialog was added which includes all legal information about Qt and its
dependencies.

Also, since the LicenseDialog not longer exists, its files were deleted.
And because the AboutDialog class was so simple, its implementation was
included into the MainWindow implementation. This reduces the number of
files, and the complexity of the project.