Replace "margin" by an "offset" field.
This commit replaces the old "margin" mechanism by a new more powerful
"offset" field. Before, when margin was used, the watermark was moved
by a few percent from its position toward the center of the image. Now,
with the new offset field, we can move the watermark in any direction,
and by any amount of pixels.