Improve crop editor rendering
Remove the useless use of trigonometric functions and of radians angles. Really `12 * 2.0 * 3.141592654 / 12` ? If simplified, we have `2pi`, `sin(2pi)=0` and `cos(2pi)=1`, allowing to completely simplify the remaining code. I really don't know why the code was so complicated, but it now partially corrected. The code was also a bit refactorised to be less verbose, avoiding for example useless variables.

Moreover, this commit implements a fallback when the crop size information text is displayed outside of the crop widget. In that case, the text is now translated to be always visible.