Use floating-point precision instead of integers for drawing
The editor drawing code was updated to use float instead of integers. This should help with High-DPI and provide better rendering quality. And if not, it does cost nothing (modern computers are largely powerful to support floating-point computations).

This commit also remove mapFrom*() and mapTo*() functions of the Editor class.