Fix zoom slider
Now the zoom slider and the zoom factor of the Editor will be connected. So when we zoom with the scroll wheel, the zoom factor will be correctly
displayed in the zoom slider.

With this commit, now the zoom slider will have
its range automatically set to the kMinZoomFactor
and kMaxZoomFactor of Editor (which are now publics).